+++ packages/cli/src/index.ts
@@ -10,7 +10,6 @@ import {
 } from '@gql2ts/util';
 import { ISchemaToInterfaceOptions, generateNamespace } from '@gql2ts/from-schema';
 import fromQuery from '@gql2ts/from-query';
-import { IFromQueryReturnValue } from '@gql2ts/types';
 
 // tslint:disable-next-line no-require-imports no-var-requires
 const { version } = require('../package.json');
@@ -21,7 +20,6 @@ program
   .option('-o --output-file [outputFile]', 'name for output file, will use stdout if not specified')
   .option('-n --namespace [namespace]', 'name for the namespace, defaults to "GQL"', 'GQL')
   .option('-i --ignored-types <ignoredTypes>', 'names of types to ignore (comma delimited)', v => v.split(','), [])
-  .option('-l --legacy', 'Use TypeScript 1.x annotation', false)
   .option('-e --external-options [externalOptions]', 'ES Module with method overwrites')
   .option('--ignore-type-name-declaration', 'Whether to exclude __typename', false)
   .option('--exclude-deprecated-fields', 'Whether to exclude deprecated fields', false)
@@ -42,12 +40,11 @@ const run: (schema: PossibleSchemaInput, options: Partial<ICLIOptions>) => void 
   if (program.args[1]) {
     const queryFile: string = program.args[1];
     const query: string = fs.readFileSync(queryFile).toString();
-    const info: IFromQueryReturnValue[] = fromQuery(schema, query, {}, defaultOverrides);
-    const toWrite: string = info.map(inf => inf.result).join('\n\n');
+    const info: string = fromQuery(schema, query, {}, defaultOverrides);
     if (options.outputFile) {
-      writeToFile(options.outputFile, toWrite);
+      writeToFile(options.outputFile, info);
     } else {
-      console.log(toWrite);
+      console.log(info);
     }
     return;
   }