+++ packages/from-query/__tests__/from-query-test.ts
@@ -1,4 +1,4 @@
-import { introspectionQuery } from 'graphql';
+// import { getIntrospectionQuery } from 'graphql';
 
 import { Query1, Query2 } from './data/query';
 import Schema from './data/schema';
@@ -14,32 +14,41 @@ describe('additional fragments', () => {
     ).toMatchSnapshot();
   });
 
-  it ('works 2', () => {
-    expect(
-      FromQuery(Schema, Query2)
-    ).toMatchSnapshot();
-  });
+  // TODO: InlineFragment Must Be Inlined!
+  // it ('works 2', () => {
+  //   expect(
+  //     FromQuery(Schema, Query2)
+  //   ).toMatchSnapshot();
+  // });
 
-  it('works with the introspection query', () => {
-    expect(
-      FromQuery(Schema, introspectionQuery)
-    ).toMatchSnapshot();
-  });
+  // TODO: Introspection not supported yet
+  // it('works with the introspection query', () => {
+  //   const query: string = getIntrospectionQuery();
 
+  //   expect(
+  //     FromQuery(Schema, query)
+  //   ).toMatchSnapshot();
+  // });
+
+  // TODO: Introspection not supported yet
   it('works with another introspective query', () => {
-    expect(
-      FromQuery(Schema, `query {
-        __type(name: "String") {
-          kind
-        }
-        t1: __type(name: "Boolean") {
-          name
-        }
-        t2: __type(name: "Float") {
-          kind
-          name
-        }
-      }`)
-    ).toMatchSnapshot();
+    const func: () => string = () =>
+      FromQuery(
+        Schema,
+        `query {
+          __type(name: "String") {
+            kind
+          }
+          t1: __type(name: "Boolean") {
+            name
+          }
+          t2: __type(name: "Float") {
+            kind
+            name
+          }
+        }`
+      );
+
+    expect(func).toThrowError('introspection not supported yet');
   });
 });